implemented remote file system in VisualEagle

+++ VisualEagle/src/main/java/org/visualeagle/gui/remotewindow/fileprovider/AbstractFileProvider.java
@@ -33,21 +33,21 @@ public abstract class AbstractFileProvider {
     /**
     Return handle to file
     */
-    public abstract ListenableFutureTask<Integer> openFileForReading(RFile file);
+    public abstract ListenableFutureTask<Long> openFileForReading(RFile file);
 
     /**
     Return handle to file
     */
-    public abstract ListenableFutureTask<Integer> openFileForWriting(RFile file, boolean append);
+    public abstract ListenableFutureTask<Long> openFileForWriting(RFile file, boolean append);
     
-    public abstract ListenableFutureTask<Boolean> writeToFile(int handle, byte[]buffer,int count);
+    public abstract ListenableFutureTask<Boolean> writeToFile(long handle, byte[]buffer,int count);
     
     /**
     Tries to fully fill buffer from file opened with @handle<br>
     returns number of actually read bytes
     */
-    public abstract ListenableFutureTask<Integer> readFromFile(int handle, byte[]buffer);
+    public abstract ListenableFutureTask<Integer> readFromFile(long handle, byte[]buffer);
     
-    public abstract ListenableFutureTask<Boolean> close(int handle);
+    public abstract ListenableFutureTask<Boolean> close(long handle);
     
 }
